fix(sworddeck): display RAM in GB instead of raw MB
- memInfo() now returns double GB values instead of int MB
- m_memUsed / m_memTotal changed to double in right_panel.h
- Display format: '6.5G / 15.3G  41%' instead of '6650M / 15657M  41%'
- Use qMax(0.001, ...) guard for the percentage division

+++ cpp-sworddeck/src/right_panel.h
@@ -49,8 +49,8 @@ private:
 
     PipesLayer m_pipes;
     double m_cpu    = 0;
-    int m_memUsed   = 0;
-    int m_memTotal  = 1;
+    double m_memUsed  = 0;
+    double m_memTotal = 1;
     int m_temp      = 0;
     bool m_hasTemp  = false;
     QList<QPair<QString, QPair<double,double>>> m_procs;
